requirements.txt
fastapi
uvicorn
pydantic-settings
pydantic>=2.7.0
x402-tron[tron,fastapi] @ git+https://github.com/open-aibank/x402-tron@v0.1.6#subdirectory=python/x402
httpx

tronpy>=0.4.0
eth-account>=0.10.0

# Database
sqlalchemy>=2.0.0
asyncpg
greenlet

# Configuration
pyyaml
onepassword-sdk

# Rate Limiting
slowapi

# Monitoring
prometheus-fastapi-instrumentator

# Testing
pytest
pytest-asyncio
pytest-mock
